Remove unused imports and misleading shebangs
A shebang tells the kernel which interpreter to use when loading a
program. `aide` and `metab_classes` are definitely not meant to be an
executable. `metab_import` could be, but the instructions state to run
it as `python metab_import.py`, suggesting that the file itself is not
an executable rather a script to be interpreted by python.