Add differential update option
Added a new command line option `--diff=<prev>` that performs a
differential update.

A differential update compares the triples produced by a run with that of
an older one (<prev>). Two distinct sets of triples are produced:
- an "Add" set which contains the set of triples that are in the current
run's triples, but not in the older sets.
- a "Sub" set which contains the set of triples that are in the older
set, but not in the current run's.

The corresponding files are written to add.nt and sub.nt.