Add metab admin form for adding and removing aliases.
Add a form which allows you to create a new alias or delete an existing
alias. Also changed the names table to extend the unique constraint to a witheld
boolean as well. This means that when you mark someone withheld you can still use
their name as an alias. Also update change user's withheld to mark all names as the same
updated withheld status.