Add prefixes to Person and Organization URIs
The prefixes are "p" for Person and "o" for Organization.

I also took the opportunity to change the code to adhere to OOP best
practices. Namely, this change ensures that the Person and Organization
constructors create objects in a good state. Before this change, the
objects' properties had to be set externally.