Fix namespace issue during photo upload
The code expected the namespace ends with a slash. My previous commit
forgot this and it caused an extra slash to be inserted preventing
photos from displaying correctly. This change adds a new Person.n_number
method that produces the portion of a Person's URI that is unique to
them. This allows a single place for changes to suffixes and prefixes.