Fix photo upload form to show middle and last name.
Fix the photo upload form to show the middle and last name in the last name box. It currently just shows the first word after a space. Changed it to show anything after the first word. There are a lot of names that don't fit the first, middle, last construct perfectly and it doesn't impact the actual functionality of uploading a photo.