Fix missing matches when middle initial in surname
The initial tool separated names based on spaces and assigned the first
string to `first_name` with the remaining text as `last_name`. This
means if "Taeber J Rapczak" wrote a paper, his name would be split as:
 * First Name: Taeber
 * Last  Name: J Rapczak

PubMed has the concept of "forename". In this example, "Taeber J" would
be the forename.

Honestly, at this point the first_name+last_name separation in the
database either needs to be cleaned up or simply combined since we only
ever match based on both parts of the name.

This change is the quick fix. It concatenates the name parts when
searching for people.