Add test cases for multiple names
A lot of cases for multiple names was missed with feature/multiname.
Also, even in the case where the number of institutes, departments, and
labs are the same, there is bad logic in that only the last institute's
ID and last department's ID are used.

For example, given institutes "UF;FSU" and departments "Chem;Bio", this
would create two institutes "UF" and "FSU" and two departments "Chem"
and "Bio", but both departments erroneously belong to "FSU".