Fetch tools from MTWiki and generate triples
Regarding the reformatting and reorg. `triples` needed to call
`prefill.parse_pubmed_authors`, but `prefill` was calling
`triples.get_config`, which would have resulted in a circular
dependency. Since `Aide` is really just an unusual name for what
appeared to be a configuration object, I renamed it and moved
`get_config` there (were it was renamed `load` as in `config.load`).

There was also so reformatting that, admittedly, I should have done in a
separate commit, but I did not.